Cleaned up the rendering of Label component.
